requirements.txt
python-telegram-bot==20.7
langchain==0.1.11
openai==1.23.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
langchain-openai==0.0.8
langchain-core==0.1.30
langchain-community==0.0.29
pytest==7.4.4
black==24.1.1
flake8==7.0.0
anthropic==0.18.1
